Adding rails validation messages in groups/form
Before this commit, if you tried to submit new group without a
description, there was a JavaScript validation that prevented the form
from being submitted, but there was no error message, so it was
confusing why nothing was happening when 'Create Group' was clicked. By
removing this piece of JS and displaying the rails validation errors,
the user will be able to see why the group isn't sucessfully being
created.